% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_D3_map.R
\name{make_D3_map}
\alias{make_D3_map}
\title{Make a D3 map}
\usage{
make_D3_map(
  countries,
  file_prefix = "D3_map",
  surrounding = TRUE,
  main = TRUE,
  margin = 0,
  resolution = "low",
  single_file = TRUE,
  attributes = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{surrounding}{\code{logical}. If \code{TRUE}, displays all countries inside the bounding box.}

\item{main}{\code{logical}. If \code{TRUE}, displays only main geographic units of a country (excludes outlying islands and territories).
For example, for the United Kingdom, if \code{main} is \code{TRUE}, only Britain, Northern Ireland,
and other nearby islands like the Isle of Man will be displayed.
If \code{main} is \code{false}, all of the territories and dependencies of the United Kingdom around the world will be displayed.}

\item{margin}{\code{numeric}. The size of the buffer between the edge of the map and the furthest extent
of the selected geometry in px. The margin applies to all four directions.}

\item{resolution}{\code{string}. Either \code{"low"} or \code{"high"}.
If \code{"low"}, uses the medium scale data (1:50m) from Natural Earth.
If \code{"high"}, uses the large scale data (1:10m) from Natural Earth.}

\item{single_file}{\code{logical}. If \code{TRUE}, the function will return a single HTML file.
This file will include all HTML, CSS, JavaScript, and GEOJSON data required to display the map in a browser.
If \code{FALSE}, the function will return an HTML, a CSS file, and a JavaScript file.
The Javascript file will include the GEOJSON data required to make the map.
The HTML file will automatically include references to the CSS and JavaScript files.
Place all of the files in the same folder and open the HTML file in a browser.}

\item{attributes}{\code{list}. An list object created by the \code{attributes()} function.
The list can also be created by hand, but must include all of the named elements
included in the list returned by the \code{attributes()} function.}

\item{width}{\code{numeric}. The width of the map in the brower in \code{px}.
If \code{width} is specified and \code{height} is \code{NULL}, the height will be calculated
automatically to display the selected countries.
If both \code{width} and \code{height} are specified, the visible area will be exanded horizontally or vertically
as necessary to achieve the requested aspect ratio. The selected countries will be centered.}

\item{height}{\code{numeric}. The height of the map in the brower in \code{px}.}

\item{file_name}{\code{string}. The file name used for any files returned by the function.
Each file name will be followed by the appropriate file extension (.html, .css, or .js).}
}
\value{
If \code{singe_file} is \code{TRUE}, creates an HTML file in the current working directory.
This file includes a HTML, CSS, JavaScript, and GEOJSON data to produce the map.
If \code{FALSE}, creates an HTML file, a CSS file, and a JavaScript file in the current working directory.
The JavaScript file will include all GEOJSON data required to display the map in a browser.
No objects are created in the R global environment.
}
\description{
This function creates an interactive D3 choropleth map of the world or of a selected group of countries.
Safari and Chrome are fully supported. Support for other browers is not guaranteed.
This function uses open source boundary data from Natural Earth (\url{https://www.naturalearthdata.com/downloads/}).
}
\examples{
make_D3_map(
width = 700, 
height = 700, 
countries = c("NOR", "ITA"), 
attributes = attributes, 
surrounding = TRUE, 
main = TRUE, 
resolution = "high", 
margin = 25, 
single_file = TRUE
)
}
